Drop restore key from CI cache
There is no need to look further than the specific cache for
the combination of os+go.sum. Either we have a hit and use the
full cache or the hash of go.sum has changed and we need a new
cache for the build. Reusing a generic cache is not useful and
can cause problems