Fix list-namespace.txt test by extending Helm's pkg/cli
_helm/pkg/cli/environment.go_ contains `settings.{namespace,config}`, and there,
the `--namespace` flag gets bound to the config,

Sadly, neither `settings.{namespace,config}` are exported, so there's no way to
bind hypper's `--namespace` flag to the config, so Helm's actions use it.

Therefore, take helm/pkg/cli/* and extend it to contain all the settings we
currently need in Hypper, plus all the settings that are specific to Hypper,
such as `noColors`, `noEmojis`.