Refactor cmd/hypper to consume Hypper's pkg/action
Stop consuming Helm's pkg/action.
Since of course, the returns from hypper/pkg/action are still helm types, we
will need to keep importing helm packages besides helm/pkg/action.

With this refactor, we make sure that we are consuming hypper/pkg/action and
that we provide a sane API for them.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>