Use relative path for bin dir in Makefile
This is to fix a case in which we build hypper on windows
based of the windows golang binary who is aware of the full
windows path, from a POSIX env like cygwin which is unaware
that is running on windows and uses a "fake" disk paths.

The result is that building the output ends up in the disk
drive that go runs out from + output path while rm tries to
clean the "fake" output path only

By using the relative path for building the binary we
avoid any mangling issues.

As a side effect this also removes the binary quotes that
were added for no discernible reason and while they didnt
affect anything AFAIK, they looked terrible.

Signed-off-by: Itxaka <igarcia@suse.com>