Add hypper lint
Adds the lint command to hypper and rules for our
annotations.

This is run *after* the normal helm rules as to
cover as much terrain as possible so on top of
the usual helm rules we can add our own rules like
in this patch.

Currently we have 3 hypper rules that are on Warning level
and cover release-name, shared-dependencies and namespace
annotations and 1 Error level rule that checks the validity
of the shared-dependencies format.

Signed-off-by: Itxaka <igarcia@suse.com>