Fix helm settings for flags
We were filling all the helm settings before parsing the
flags, which means that anything using those helm settings for
any action would just get the default values instead of any cli
values.

This fixes the issue by creating a fill function and calling
it after the flags have been parsed, thus making sure that those
flags are properly propagated.

Otherwise the envsettings behavior its the same as it was before
as the default hypoper values are propagated to helm settings on
Envsettings construct

Signed-off-by: Itxaka <igarcia@suse.com>