feat: Padded output for when installing shared-deps
Add prefix "- " to install output and pad it to the number of recursions.
It will be ignored on first call (direct install, not shared-dep), and
it will nest itself when installing shared-deps of shared-deps.

Implemented by adding a `lvl int` var to recursive functions.

Example:

```console
$ ./bin/hypper install our-app --create-namespace
Installing shared dependencies for chart "our-app":
- Installing chart "postgresql" as "postgresql" in namespace "hypper"…
- Installing chart "prometheus" as "prometheus" in namespace "hypper"…
Installing chart "our-app" as "our-app-name" in namespace "hypper"…
Done! 👏
$ ./bin/hypper install our-app --create-namespace
Installing shared dependencies for chart "our-app":
- Shared dependency chart "postgresql" already installed, skipping
- Shared dependency chart "prometheus" already installed, skipping
Installing chart "our-app" as "our-app-name" in namespace "hypper"…
Error: cannot re-use a name that is still in use
```

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>