fix: Honour dependency repos using `file://`
Works for absolute paths (e.g: `file:///home/vic/hypper/chart`)
and for relative paths.

NOTE:

Relative paths are relative to the working directory where the Hypper
executable was run; not relative to the directory of the parent chart.

This is a different behaviour than `file://` repositories used in
"local" Helm chart dependencies, which are relative to their parent
`chart/` folder.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>