feat: Make `repository: file://` relative to parent chart
For that, add a new field Package.ParentChartPath, which contains the
abs path of parent chart, if existing.

Pass the ParentChartPath from cmd/hypper/install to
pkg/action/install.Run(), which now depends on knowing an abs path for
the chart it needs to install.

Use that chart path in pkg/action/install.LoadChart(), when the repo has
prefix `file://`, to obtain the correct relative local path.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>