Fixing db not looking at repos
When f19146a15ac4c750efc2e05ac988b7c04bc598d6 landed the repo
location was switched for the registry location. The registry is
used with OCI. This was likely an accidental autocomplete name
change. It caused the solver to be unaware of the repos and charts
within them. Changing to the right location fixes the issue.

Closes #181

Signed-off-by: Matt Farina <matt.farina@suse.com>