Fixing issue where charts in repos not erroring on dep issues
The situation is that you have A > B > C. The repo containing B and
C are missing. The code to load files was loading B but not C.
The solver would silently skip C and just install A and B.

For security reasons, we should not install charts from repos
when the repo has not been explicitly added. This is out other
system package managers work and is security guidance out of
reviews of Helm.

The solution here was to look for local files and then throw
an error if a depenency isn't a local file or in a repo.

Closes #182

Signed-off-by: Matt Farina <matt.farina@suse.com>